pyproject.toml
[project]
name = "comfyui-digit"
version = "3.0.0"
description = "DIGIT — Deep Infrastructure for Generative Image Training. ComfyUI nodes for Gemini generation, Veo video, LoRA training, dataset management, and auto-captioning."
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "requests",
    "google-auth",
    "google-genai",
    "google-cloud-storage",
    "piexif",
    "opencv-python",
    "pyyaml",
]

[project.optional-dependencies]
training = [
    "torch>=2.1.0",
    "diffusers>=0.30.0",
    "transformers>=4.40.0",
    "peft>=0.12.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    "tensorboard",
    "wandb",
    "safetensors",
]

[project.urls]
Repository = "https://github.com/thedepartmentofexternalservices/comfyui-digit"

[tool.comfy]
PublisherId = "digit"
DisplayName = "DIGIT Nodes"
